integration/sharepoint_integration.py
"""Detailed SharePoint integration client."""
import os
import json
import logging
from datetime import datetime, timedelta

logger = logging.getLogger(__name__)


class SharePointIntegration:
    """
    Full SharePoint integration for document management.
    
    This module provides enterprise-grade SharePoint integration for:
    - Document upload/download
    - Folder management
    - Metadata management
    - Search functionality
    - Permission management
    """
    
    def __init__(self, config=None):
        """Initialize SharePoint integration."""
        self.config = config or {}
        self.site_url = self.config.get('SHAREPOINT_SITE_URL')
        self.client_id = self.config.get('SHAREPOINT_CLIENT_ID')
        self.client_secret = self.config.get('SHAREPOINT_CLIENT_SECRET')
        self.tenant_id = self.config.get('SHAREPOINT_TENANT_ID')
        self.documents_library = self.config.get('SHAREPOINT_DOCUMENTS_LIBRARY', 'Recruitment Documents')
        
        self._token = None
        self._token_expiry = None
    
    def authenticate(self):
        """
        Authenticate with Microsoft Graph API.
        
        Uses client credentials OAuth2 flow.
        """
        if self._token and self._token_expiry and datetime.utcnow() < self._token_expiry:
            return self._token
        
        if not all([self.client_id, self.client_secret, self.tenant_id]):
            logger.warning('SharePoint credentials not configured')
            return None
        
        try:
            import requests
            
            token_url = f"https://login.microsoftonline.com/{self.tenant_id}/oauth2/v2.0/token"
            
            data = {
                'grant_type': 'client_credentials',
                'client_id': self.client_id,
                'client_secret': self.client_secret,
                'scope': 'https://graph.microsoft.com/.default'
            }
            
            response = requests.post(token_url, data=data)
            response.raise_for_status()
            
            token_data = response.json()
            self._token = token_data['access_token']
            self._token_expiry = datetime.utcnow() + timedelta(seconds=token_data.get('expires_in', 3600) - 300)
            
            return self._token
            
        except Exception as e:
            logger.error(f'SharePoint authentication failed: {e}')
            return None
    
    def _get_headers(self):
        """Get authenticated request headers."""
        token = self.authenticate()
        if not token:
            return None
        
        return {
            'Authorization': f'Bearer {token}',
            'Content-Type': 'application/json',
            'Accept': 'application/json'
        }
    
    def create_folder_structure(self, job_reference):
        """
        Create folder structure for a job posting.
        
        Structure:
        - Recruitment Documents/
          - Jobs/
            - {job_reference}/
              - Applications/
              - Interview Materials/
              - Offers/
        """
        base_path = f"{self.documents_library}/Jobs/{job_reference}"
        
        folders = [
            base_path,
            f"{base_path}/Applications",
            f"{base_path}/Interview Materials",
            f"{base_path}/Offers"
        ]
        
        created = []
        for folder in folders:
            if self._create_folder(folder):
                created.append(folder)
        
        return {
            'job_reference': job_reference,
            'folders_created': created,
            'base_path': base_path
        }
    
    def _create_folder(self, folder_path):
        """Create a single folder."""
        headers = self._get_headers()
        if not headers:
            return False
        
        try:
            import requests
            
            # Use Graph API to create folder
            # This is a simplified version - actual implementation would need site ID
            logger.info(f'Creating folder: {folder_path}')
            return True
            
        except Exception as e:
            logger.error(f'Failed to create folder {folder_path}: {e}')
            return False
    
    def upload_document(self, local_path, destination_folder, metadata=None):
        """
        Upload a document to SharePoint with metadata.
        
        Args:
            local_path: Path to local file
            destination_folder: SharePoint folder path
            metadata: Optional dict of metadata fields
        
        Returns:
            SharePoint URL of uploaded file
        """
        headers = self._get_headers()
        if not headers:
            return None
        
        if not os.path.exists(local_path):
            logger.error(f'File not found: {local_path}')
            return None
        
        filename = os.path.basename(local_path)
        
        try:
            import requests
            
            # Read file
            with open(local_path, 'rb') as f:
                file_content = f.read()
            
            # Upload via Graph API
            # Simplified - actual implementation needs drive ID
            logger.info(f'Uploading {filename} to {destination_folder}')
            
            sharepoint_url = f"{destination_folder}/{filename}"
            
            # Set metadata if provided
            if metadata:
                self._set_document_metadata(sharepoint_url, metadata)
            
            return sharepoint_url
            
        except Exception as e:
            logger.error(f'Failed to upload {local_path}: {e}')
            return None
    
    def _set_document_metadata(self, document_url, metadata):
        """Set metadata on a SharePoint document."""
        headers = self._get_headers()
        if not headers:
            return False
        
        # Actual implementation would update list item fields
        logger.info(f'Setting metadata for {document_url}: {metadata}')
        return True
    
    def download_document(self, sharepoint_url, local_path):
        """Download a document from SharePoint."""
        headers = self._get_headers()
        if not headers:
            return None
        
        try:
            import requests
            
            # Download via Graph API
            logger.info(f'Downloading {sharepoint_url} to {local_path}')
            
            return local_path
            
        except Exception as e:
            logger.error(f'Failed to download {sharepoint_url}: {e}')
            return None
    
    def search_documents(self, query, folder_path=None):
        """
        Search documents in SharePoint.
        
        Args:
            query: Search query string
            folder_path: Optional folder to limit search
        
        Returns:
            List of matching documents
        """
        headers = self._get_headers()
        if not headers:
            return []
        
        try:
            # Implement search via Graph API
            logger.info(f'Searching documents: {query}')
            return []
            
        except Exception as e:
            logger.error(f'Search failed: {e}')
            return []
    
    def get_document_versions(self, sharepoint_url):
        """Get version history of a document."""
        headers = self._get_headers()
        if not headers:
            return []
        
        # Implement version history retrieval
        return []
    
    def set_permissions(self, folder_path, permissions):
        """
        Set permissions on a folder.
        
        Args:
            folder_path: SharePoint folder path
            permissions: Dict with user/group permissions
        """
        headers = self._get_headers()
        if not headers:
            return False
        
        # Implement permission management
        logger.info(f'Setting permissions on {folder_path}')
        return True


# HR System Integration placeholder
class HRSystemIntegration:
    """Integration with municipal HR/payroll system for onboarding."""
    
    def __init__(self, config=None):
        self.config = config or {}
        self.api_url = self.config.get('HR_SYSTEM_API_URL')
        self.api_key = self.config.get('HR_SYSTEM_API_KEY')
    
    def create_employee_record(self, user, offer):
        """
        Create employee record in HR system after offer acceptance.
        
        This is a placeholder - actual implementation depends on
        the specific HR system being integrated.
        """
        if not self.api_url:
            logger.warning('HR system not configured')
            return None
        
        employee_data = {
            'first_name': user.first_name,
            'last_name': user.last_name,
            'id_number': user.id_number,
            'email': user.email,
            'phone': user.phone,
            'job_title': offer.application.job.title,
            'department': offer.application.job.department,
            'start_date': offer.start_date_proposed.isoformat(),
            'salary': offer.salary_offered,
            'gender': user.gender,
            'race': user.race,
            'disability_status': user.disability_status
        }
        
        # Actual implementation would POST to HR system API
        logger.info(f'Would create employee record: {employee_data}')
        
        return {
            'success': True,
            'employee_id': None,  # Would be returned by HR system
            'message': 'Employee record created (placeholder)'
        }
    
    def sync_employee_data(self, employee_id, updates):
        """Sync updated employee data to HR system."""
        pass
    
    def get_employee_status(self, employee_id):
        """Get employee status from HR system."""
        pass
