requirements.txt
# E-Recruitment Portal - Requirements

# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1

# Database
PyMySQL==1.1.0
cryptography==41.0.7

# Background tasks
celery==5.3.4
redis==5.0.1

# Forms and validation
WTForms==3.1.1
email-validator==2.1.0

# Security
Werkzeug==3.0.1
python-dotenv==1.0.0

# SharePoint integration
Office365-REST-Python-Client==2.5.2

# Utilities
python-dateutil==2.8.2
Pillow==10.1.0

# For SMS (optional)
# twilio==8.10.0

# Development
pytest==7.4.3
pytest-flask==1.3.0

# Production WSGI server
gunicorn==21.2.0
